+++ src/django/oar/urls.py
@@ -35,6 +35,9 @@ router.register('facility-claims', views.FacilityClaimViewSet,
 router.register('facility-matches', views.FacilityMatchViewSet,
                 'facility-match')
 router.register('api-blocks', views.ApiBlockViewSet, 'api-block')
+router.register('facility-activity-reports',
+                views.FacilityActivityReportViewSet,
+                'facility-activity-report')
 
 public_apis = [
     url(r'^api/', include(router.urls)),